move instantiateComputedArtifacts to Runner
also removes `requestundefined()` from ComputedArtifact base class
