+++ lighthouse-core/test/audits/manifest-short-name-length-test.js
@@ -22,10 +22,10 @@ const manifestParser = require('../../lib/manifest-parser');
 const EXAMPLE_MANIFEST_URL = 'https://example.com/manifest.json';
 const EXAMPLE_DOC_URL = 'https://example.com/index.html';
 
-const GatherRunner = require('../../gather/gather-runner.js');
+const Runner = require('../../runner.js');
 
 function generateMockArtifacts() {
-  const computedArtifacts = GatherRunner.instantiateComputedArtifacts();
+  const computedArtifacts = Runner.instantiateComputedArtifacts();
   const mockArtifacts = Object.assign({}, computedArtifacts, {
     Manifest: null
   });
